:recycle: refactor(ProjectEndpointTest): Moved rejection handler related tests into a dedicated class
